fix: organized per suggestions
