+++ src/pages/Houses.js
@@ -1,14 +1,11 @@
-// import '../styles/Houses.css';
 import React, {useState, useEffect} from 'react';
-import ReactDOM from 'react-dom/client';
+
+import '../styles/Houses.css';
 
 const Houses = () => {
     return (
         <>
-            <br />
-            <br />
-            <br />
-            <h1>Houses</h1>;
+            <h1 className="Houses">Houses</h1>;
         </>
     );
 };
