+++ src/styles/Books.css
@@ -1,0 +1,4 @@
+.Books {
+    padding: 100px;
+    text-align: center;
+}
