+++ src/styles/Characters.css
@@ -1,0 +1,5 @@
+.Characters {
+    padding: 100px;
+    text-align: center;
+    color: blue;
+}
