+++ src/styles/FrontPage.css
@@ -1,5 +1,5 @@
 .FrontPage {
-    background-image: url('../assets/frontpage.png');
+    background-image: url('../assets/images/frontpage.png');
     height: 100vh;
     width: 100vw;
     /* background-size: cover; */
