+++ src/styles/Houses.css
@@ -1,0 +1,4 @@
+.Houses {
+    padding: 100px;
+    text-align: center;
+}
