+++ src/styles/Login.css
@@ -1,7 +1,7 @@
 .Login-Wrapper {
     height: 100vh;
     background-color: rgb(0, 0, 0);
-    background-image: url('../assets/frontpage.png');
+    background-image: url('../assets/images/frontpage.png');
 }
 
 .Login {
