+++ src/styles/Signup.css
@@ -1,7 +1,7 @@
 .Signup-Wrapper {
     height: 100vh;
     background-color: rgb(0, 0, 0);
-    background-image: url('../assets/frontpage.png');
+    background-image: url('../assets/images/frontpage.png');
 }
 
 .Signup {
